refactor(vmm): split manager core
