+++ virt/arcbox-vmm/src/vmm/mod.rs
@@ -9,7 +9,6 @@
 
 #[cfg(target_os = "macos")]
 use std::os::fd::{AsRawFd, FromRawFd, OwnedFd};
-use std::path::PathBuf;
 use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::time::Duration;
@@ -17,13 +16,10 @@ use std::time::Duration;
 use crate::device::DeviceManager;
 use crate::error::{Result, VmmError};
 use crate::event::EventLoop;
-use crate::irq::{Gsi, IrqChip, IrqTriggerCallback};
+use crate::irq::IrqChip;
 use crate::memory::MemoryManager;
 #[cfg(target_os = "linux")]
 use crate::vcpu::VcpuManager;
-
-use arcbox_hypervisor::VirtioDeviceConfig;
-use arcbox_hypervisor::VmConfig;
 
 #[cfg(target_os = "macos")]
 mod darwin;
@@ -32,154 +28,16 @@ mod darwin_hv;
 #[cfg(target_os = "linux")]
 mod linux;
 
-/// Shared directory configuration for `VirtioFS`.
-#[derive(Debug, Clone)]
-pub struct SharedDirConfig {
-    /// Host path to share.
-    pub host_path: PathBuf,
-    /// Tag for mounting in guest.
-    pub tag: String,
-    /// Whether the share is read-only.
-    pub read_only: bool,
-}
+mod config;
+mod event_handlers;
+mod helpers;
+#[cfg(test)]
+mod tests;
 
-/// Block device configuration for `VirtIO` block devices.
-#[derive(Debug, Clone)]
-pub struct BlockDeviceConfig {
-    /// Path to the disk image file on the host.
-    pub path: PathBuf,
-    /// Whether the block device is read-only.
-    pub read_only: bool,
-}
-
-/// VMM-specific configuration.
-#[derive(Debug, Clone)]
-pub struct VmmConfig {
-    /// Number of virtual CPUs.
-    pub vcpu_count: u32,
-    /// Memory size in bytes.
-    pub memory_size: u64,
-    /// Path to the kernel image.
-    pub kernel_path: PathBuf,
-    /// Kernel command line arguments.
-    pub kernel_cmdline: String,
-    /// Path to initial ramdisk (optional).
-    pub initrd_path: Option<PathBuf>,
-    /// Enable Rosetta 2 translation (macOS ARM only).
-    pub enable_rosetta: bool,
-    /// Enable serial console.
-    pub serial_console: bool,
-    /// Enable virtio-console.
-    pub virtio_console: bool,
-    /// Shared directories for `VirtioFS`.
-    pub shared_dirs: Vec<SharedDirConfig>,
-    /// Enable networking.
-    pub networking: bool,
-    /// Enable vsock.
-    pub vsock: bool,
-    /// Guest CID for vsock connections (Linux).
-    pub guest_cid: Option<u32>,
-    /// Enable memory balloon device.
-    ///
-    /// The balloon device allows dynamic memory management by inflating
-    /// (reclaiming memory from guest) or deflating (returning memory).
-    /// This helps achieve low idle memory usage.
-    pub balloon: bool,
-    /// Block devices to attach to the VM.
-    pub block_devices: Vec<BlockDeviceConfig>,
-    /// Optional MAC address for the bridge NAT NIC on macOS.
-    pub bridge_nic_mac: Option<String>,
-    /// VM backend selection (macOS only).
-    ///
-    /// Controls whether to use Hypervisor.framework (custom VMM with TSO
-    /// support) or Virtualization.framework (managed execution). Requires
-    /// macOS 15+ for HV backend. See [`VmBackend`] for details.
-    pub backend: VmBackend,
-}
-
-impl Default for VmmConfig {
-    fn default() -> Self {
-        Self {
-            vcpu_count: arcbox_hypervisor::default_vm_cpu_count(),
-            memory_size: arcbox_hypervisor::default_vm_memory_size(),
-            kernel_path: PathBuf::new(),
-            kernel_cmdline: String::new(),
-            initrd_path: None,
-            enable_rosetta: false,
-            serial_console: true,
-            virtio_console: true,
-            shared_dirs: Vec::new(),
-            networking: true,
-            vsock: true,
-            guest_cid: None,
-            balloon: true, // Enable balloon by default for memory optimization
-            block_devices: Vec::new(),
-            bridge_nic_mac: None,
-            backend: VmBackend::default(),
-        }
-    }
-}
-
-impl VmmConfig {
-    /// Creates a `VmConfig` for the hypervisor from this VMM config.
-    fn to_vm_config(&self) -> VmConfig {
-        let mut builder = VmConfig::builder()
-            .vcpu_count(self.vcpu_count)
-            .memory_size(self.memory_size)
-            .kernel_path(self.kernel_path.to_string_lossy())
-            .kernel_cmdline(&self.kernel_cmdline)
-            .enable_rosetta(self.enable_rosetta);
-
-        if let Some(initrd_path) = &self.initrd_path {
-            builder = builder.initrd_path(initrd_path.to_string_lossy());
-        }
-
-        builder.build()
-    }
-}
-
-/// VM backend selection for macOS.
-///
-/// Controls whether the VMM uses Apple's Virtualization.framework (managed
-/// execution) or Hypervisor.framework (custom VMM with manual execution).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub enum VmBackend {
-    /// Auto-select: HV for native ARM64, VZ when Rosetta is needed.
-    #[default]
-    Auto,
-    /// Force Hypervisor.framework (custom VMM). Requires macOS 15+.
-    Hv,
-    /// Force Virtualization.framework (managed execution).
-    Vz,
-}
-
-/// Resolved backend after evaluating platform constraints.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ResolvedBackend {
-    /// Hypervisor.framework (custom VMM with manual vCPU execution).
-    Hv,
-    /// Virtualization.framework (managed execution by Apple's runtime).
-    Vz,
-}
-
-/// VMM state.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum VmmState {
-    /// VMM created but not started.
-    Created,
-    /// VMM is initializing.
-    Initializing,
-    /// VMM is running.
-    Running,
-    /// VMM is paused.
-    Paused,
-    /// VMM is stopping.
-    Stopping,
-    /// VMM is stopped.
-    Stopped,
-    /// VMM encountered an error.
-    Failed,
-}
+pub use config::{
+    BlockDeviceConfig, ResolvedBackend, SharedDirConfig, VmBackend, VmmConfig, VmmState,
+};
+use helpers::{placeholder_vcpu_snapshots, resolve_backend};
 
 /// Virtual Machine Monitor.
 ///
@@ -871,319 +729,6 @@ impl Vmm {
         tracing::info!("VMM exited");
         Ok(())
     }
-
-    /// Handles an event from the event loop.
-    fn handle_event(&self, event: crate::event::VmmEvent) {
-        use crate::event::VmmEvent;
-
-        match event {
-            VmmEvent::VcpuExit { vcpu_id, exit } => {
-                self.handle_vcpu_exit(vcpu_id, exit);
-            }
-            VmmEvent::DeviceIo {
-                device_id,
-                is_read,
-                addr,
-                data,
-            } => {
-                self.handle_device_io(device_id, is_read, addr, data);
-            }
-            VmmEvent::Timer { id } => {
-                tracing::trace!("Timer {} fired", id);
-                // Timer handling would go here (e.g., for RTC, PIT, etc.)
-            }
-            VmmEvent::Shutdown => {
-                tracing::info!("Shutdown requested");
-                self.running.store(false, Ordering::SeqCst);
-            }
-        }
-    }
-
-    /// Handles a vCPU exit event.
-    ///
-    /// This processes exits from the hypervisor such as I/O, MMIO, and special
-    /// instructions that require VMM intervention.
-    fn handle_vcpu_exit(&self, vcpu_id: u32, exit: arcbox_hypervisor::VcpuExit) {
-        use arcbox_hypervisor::VcpuExit;
-
-        match &exit {
-            VcpuExit::Halt => {
-                tracing::debug!("vCPU {} halted", vcpu_id);
-                // HLT instruction - guest is idle, can reduce CPU usage
-                // In a real implementation, we might pause the vCPU until an interrupt
-            }
-
-            VcpuExit::IoOut { port, size, data } => {
-                tracing::trace!(
-                    "vCPU {} I/O out: port={:#x}, size={}, data={:#x}",
-                    vcpu_id,
-                    port,
-                    size,
-                    data
-                );
-                self.handle_io_out(*port, *size, *data);
-            }
-
-            VcpuExit::IoIn { port, size } => {
-                tracing::trace!("vCPU {} I/O in: port={:#x}, size={}", vcpu_id, port, size);
-                let _value = self.handle_io_in(*port, *size);
-                // Note: The value would need to be written back to the vCPU,
-                // which requires access to the vCPU registers.
-            }
-
-            VcpuExit::MmioRead { addr, size } => {
-                tracing::trace!(
-                    "vCPU {} MMIO read: addr={:#x}, size={}",
-                    vcpu_id,
-                    addr,
-                    size
-                );
-                if let Some(ref device_manager) = self.device_manager {
-                    match device_manager.handle_mmio_read(*addr, *size as usize) {
-                        Ok(value) => {
-                            tracing::trace!("MMIO read returned: {:#x}", value);
-                            // Note: Value would need to be written back to vCPU
-                        }
-                        Err(e) => {
-                            tracing::warn!("MMIO read failed at {:#x}: {}", addr, e);
-                        }
-                    }
-                }
-            }
-
-            VcpuExit::MmioWrite { addr, size, data } => {
-                tracing::trace!(
-                    "vCPU {} MMIO write: addr={:#x}, size={}, data={:#x}",
-                    vcpu_id,
-                    addr,
-                    size,
-                    data
-                );
-                if let Some(ref device_manager) = self.device_manager {
-                    if let Err(e) = device_manager.handle_mmio_write(*addr, *size as usize, *data) {
-                        tracing::warn!("MMIO write failed at {:#x}: {}", addr, e);
-                    }
-                }
-            }
-
-            VcpuExit::Hypercall { nr, args } => {
-                tracing::debug!("vCPU {} hypercall: nr={}, args={:?}", vcpu_id, nr, args);
-                // Hypercall handling - used for paravirtualization
-                self.handle_hypercall(vcpu_id, *nr, *args);
-            }
-
-            VcpuExit::SystemReset => {
-                tracing::info!("vCPU {} requested system reset", vcpu_id);
-                // Guest requested a reset - could restart VM or signal caller
-                self.running.store(false, Ordering::SeqCst);
-            }
-
-            VcpuExit::Shutdown => {
-                tracing::info!("vCPU {} requested shutdown", vcpu_id);
-                self.running.store(false, Ordering::SeqCst);
-            }
-
-            VcpuExit::Unknown(code) => {
-                tracing::warn!("vCPU {} unknown exit: {}", vcpu_id, code);
-            }
-
-            VcpuExit::Debug => {
-                tracing::debug!("vCPU {} unhandled exit: {:?}", vcpu_id, exit);
-            }
-        }
-    }
-
-    /// Handles device I/O events.
-    fn handle_device_io(&self, device_id: u32, is_read: bool, addr: u64, data: Option<u64>) {
-        if let Some(ref device_manager) = self.device_manager {
-            if is_read {
-                match device_manager.handle_mmio_read(addr, 4) {
-                    Ok(value) => {
-                        tracing::trace!("Device {} read at {:#x}: {:#x}", device_id, addr, value);
-                    }
-                    Err(e) => {
-                        tracing::warn!("Device {} read failed: {}", device_id, e);
-                    }
-                }
-            } else if let Some(value) = data {
-                if let Err(e) = device_manager.handle_mmio_write(addr, 4, value) {
-                    tracing::warn!("Device {} write failed: {}", device_id, e);
-                }
-            }
-        }
-    }
-
-    /// Handles I/O port output.
-    fn handle_io_out(&self, port: u16, size: u8, data: u64) {
-        match port {
-            // Serial ports (COM1-COM4)
-            0x3F8..=0x3FF => {
-                // COM1 - primary serial port
-                if port == 0x3F8 {
-                    // Data register - output character
-                    let ch = (data & 0xFF) as u8;
-                    if ch.is_ascii() && (ch.is_ascii_graphic() || ch.is_ascii_whitespace()) {
-                        tracing::trace!("Serial output: '{}'", ch as char);
-                    }
-                }
-            }
-            0x2F8..=0x2FF => {
-                // COM2
-            }
-
-            // Debug port (Bochs/QEMU convention)
-            0x402 => {
-                let ch = (data & 0xFF) as u8;
-                if ch.is_ascii() {
-                    tracing::debug!("Debug port: '{}'", ch as char);
-                }
-            }
-
-            // ACPI power management
-            0x604 => {
-                // ACPI PM1a control - check for S5 (shutdown)
-                if data & 0x2000 != 0 {
-                    tracing::info!("ACPI shutdown requested");
-                    // Would signal shutdown here
-                }
-            }
-
-            // PIC (Programmable Interrupt Controller)
-            0x20 | 0x21 | 0xA0 | 0xA1 => {
-                tracing::trace!("PIC write: port={:#x}, data={:#x}", port, data);
-            }
-
-            // PIT (Programmable Interval Timer)
-            0x40..=0x43 => {
-                tracing::trace!("PIT write: port={:#x}, data={:#x}", port, data);
-            }
-
-            _ => {
-                tracing::trace!(
-                    "Unhandled I/O out: port={:#x}, size={}, data={:#x}",
-                    port,
-                    size,
-                    data
-                );
-            }
-        }
-    }
-
-    /// Handles I/O port input.
-    fn handle_io_in(&self, port: u16, size: u8) -> u64 {
-        match port {
-            // Serial ports - Line Status Register
-            0x3FD => {
-                // LSR: Always report transmitter empty (0x60)
-                0x60
-            }
-
-            // RTC (Real-Time Clock)
-            0x70 | 0x71 => 0,
-
-            // Keyboard controller status
-            0x64 => {
-                // Report output buffer empty, input buffer not full
-                0x00
-            }
-
-            // PIC
-            0x20 | 0x21 | 0xA0 | 0xA1 => 0xFF,
-
-            _ => {
-                tracing::trace!("Unhandled I/O in: port={:#x}, size={}", port, size);
-                0xFF // Return all 1s for unhandled ports
-            }
-        }
-    }
-
-    /// Handles a hypercall from the guest.
-    fn handle_hypercall(&self, vcpu_id: u32, nr: u64, args: [u64; 6]) {
-        // Common hypercall numbers (KVM convention):
-        // 0: KVM_HC_VAPIC_POLL_IRQ
-        // 1: KVM_HC_MMU_OP (deprecated)
-        // 2: KVM_HC_FEATURES
-        // 3: KVM_HC_PPC_MAP_MAGIC_PAGE
-        // 4: KVM_HC_KICK_CPU
-        // 5: KVM_HC_SEND_IPI
-        // 9: KVM_HC_MAP_GPA_RANGE
-
-        match nr {
-            0 => {
-                // Poll for pending interrupts
-                tracing::trace!("vCPU {} hypercall: VAPIC_POLL_IRQ", vcpu_id);
-            }
-            2 => {
-                // Get features
-                tracing::trace!("vCPU {} hypercall: GET_FEATURES", vcpu_id);
-            }
-            4 => {
-                // Kick another vCPU
-                let target_vcpu = args[0] as u32;
-                tracing::trace!(
-                    "vCPU {} hypercall: KICK_CPU target={}",
-                    vcpu_id,
-                    target_vcpu
-                );
-            }
-            _ => {
-                tracing::debug!(
-                    "vCPU {} unhandled hypercall: nr={}, args={:?}",
-                    vcpu_id,
-                    nr,
-                    args
-                );
-            }
-        }
-    }
-}
-
-/// Resolves the backend selection based on platform constraints.
-///
-/// When `Auto` is selected, Rosetta requires VZ (Hypervisor.framework cannot
-/// translate x86_64 instructions). Otherwise, default to VZ until the HV
-/// backend is fully validated.
-#[cfg(target_os = "macos")]
-fn resolve_backend(config: &VmmConfig) -> ResolvedBackend {
-    match config.backend {
-        VmBackend::Vz => ResolvedBackend::Vz,
-        VmBackend::Hv => ResolvedBackend::Hv,
-        // Auto: Rosetta x86_64 translation requires VZ (Hypervisor.framework
-        // cannot translate instructions). For native ARM64 workloads, prefer HV.
-        VmBackend::Auto => {
-            if config.enable_rosetta {
-                ResolvedBackend::Vz
-            } else {
-                ResolvedBackend::Hv
-            }
-        }
-    }
-}
-
-fn placeholder_vcpu_snapshots(vcpu_count: u32) -> Vec<arcbox_hypervisor::VcpuSnapshot> {
-    #[cfg(target_arch = "aarch64")]
-    {
-        (0..vcpu_count)
-            .map(|id| {
-                arcbox_hypervisor::VcpuSnapshot::new_arm64(
-                    id,
-                    arcbox_hypervisor::Arm64Registers::default(),
-                )
-            })
-            .collect()
-    }
-
-    #[cfg(not(target_arch = "aarch64"))]
-    {
-        (0..vcpu_count)
-            .map(|id| {
-                arcbox_hypervisor::VcpuSnapshot::new_x86(
-                    id,
-                    arcbox_hypervisor::Registers::default(),
-                )
-            })
-            .collect()
-    }
 }
 
 impl Vmm {
@@ -1211,96 +756,6 @@ impl Drop for Vmm {
             } else {
                 let _ = self.stop();
             }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_vmm_creation() {
-        let config = VmmConfig {
-            guest_cid: Some(3),
-            ..Default::default()
-        };
-        let vmm = Vmm::new(config).unwrap();
-        assert_eq!(vmm.state(), VmmState::Created);
-    }
-
-    #[test]
-    fn test_vmm_invalid_config() {
-        // Zero vCPUs
-        let config = VmmConfig {
-            vcpu_count: 0,
-            guest_cid: Some(3),
-            ..Default::default()
-        };
-        assert!(Vmm::new(config).is_err());
-
-        // Too little memory
-        let config = VmmConfig {
-            memory_size: 1024, // 1KB
-            guest_cid: Some(3),
-            ..Default::default()
-        };
-        assert!(Vmm::new(config).is_err());
-    }
-
-    #[test]
-    fn test_vmm_requires_guest_cid_when_vsock_enabled() {
-        let config = VmmConfig {
-            guest_cid: None,
-            ..Default::default()
-        };
-        assert!(Vmm::new(config).is_err());
-
-        let config = VmmConfig {
-            vsock: false,
-            guest_cid: None,
-            ..Default::default()
-        };
-        assert!(Vmm::new(config).is_ok());
-    }
-
-    #[test]
-    fn test_vmm_state_transitions() {
-        let config = VmmConfig {
-            guest_cid: Some(3),
-            ..Default::default()
-        };
-        let mut vmm = Vmm::new(config).unwrap();
-
-        // Can't pause before running
-        assert!(vmm.pause().is_err());
-
-        // Can't resume before pausing
-        assert!(vmm.resume().is_err());
-    }
-
-    /// With the HV backend resolved, `capture_snapshot_context` must return
-    /// `VmmError::Unsupported` rather than the old generic `invalid_state
-    /// ("hypervisor VM handle is unavailable")` error. This covers the
-    /// ABX-360 correctness gap: HV-backed VMs were silently hitting
-    /// VZ-only snapshot code paths.
-    #[cfg(target_os = "macos")]
-    #[test]
-    fn test_hv_snapshot_returns_unsupported() {
-        let config = VmmConfig {
-            guest_cid: Some(3),
-            ..Default::default()
-        };
-        let mut vmm = Vmm::new(config).unwrap();
-
-        // Simulate a started HV VM without actually booting one.
-        vmm.resolved_backend = Some(ResolvedBackend::Hv);
-        vmm.state = VmmState::Running;
-
-        match vmm.capture_snapshot_context() {
-            Err(VmmError::Unsupported(msg)) => assert!(msg.contains("HV backend")),
-            Err(other) => panic!("expected Unsupported for HV capture, got Err({other:?})"),
-            Ok(_) => panic!("expected Unsupported for HV capture, got Ok"),
         }
     }
 }
